fixed some flaws

+++ src/homework2/task7.py
@@ -1,9 +1,9 @@
 # Check existence of a triangle. Count square
 
 
-a = int(input('Enter side a '))
-b = int(input('Enter side b '))
-c = int(input('Enter side c '))
+a = float(input('Enter side a '))
+b = float(input('Enter side b '))
+c = float(input('Enter side c '))
 if a < b + c and b < a + c and c < a + b:
     print('Triangle with sides', a, b, c, 'exists')
     p = (a + b + c) / 2
